refactor: organize components into modules and adopt kebab-case
- Move components into a modular directory structure under `src/components`.
- Rename all component files from PascalCase to kebab-case.
- Update all import paths to reflect the new file locations and naming.
- Remove `index.ts` barrel files in favor of direct imports.
