+++ src/components/common/CopyButton.tsx
@@ -1,6 +1,6 @@
 import { useState } from 'react'
 import { Button } from 'react-aria-components'
-import { queue } from '../ToastQueue'
+import { queue } from '@/components/toasts/toast-queue'
 import { getErrorMessage } from '../../helpers/getErrorMessage'
 import { CheckIcon, CopyIcon } from '@radix-ui/react-icons'
