+++ src/components/sidebar/tree-node-item.tsx
