+++ src/main.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 import ReactDOM from 'react-dom/client'
 import './main.css'
 import App from './App'
-import { Toaster } from './components/ToastQueue'
+import { Toaster } from './components/toasts/toast-queue'
 import { Provider } from 'jotai'
 
 ReactDOM.createRoot(document.getElementById('root') as HTMLElement).render(
